Preserve authentication as we navigate the API
The client accepts authentication information in the API entry-point URL, but
failed to preserve that as it navigated through.  This resulted in an
immediate failure if authentication was enabled -- as soon as you tried to do
anything other than view available commands, authentication was forgotten, and
you got a nasty 401 response.

This fixes that, and adds tests to ensure we get things right; we simply copy
the authentication data from the previous request, if any, into the
next request.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>