Improve "Unauthorized" error message
When authentication is enabled on the server and credentials are not supplied on the client, the error message should advise the user to supply credentials. This previously resulted in a misleading error that it could not connect to the server. Also reformatted error strings for legibility.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-175