Report HTTP errors to the user
This change actually reports HTTP errors to the user, including the HTTP body,
when the server fails.  This, in turn, allows the server to actually
communicate information to the user about why things went wrong.

It also adjusts the wording of the default message to be clearer about the
"failure" -- that communication failed, not necessarily just connectivity
failed with the server.

https://tickets.puppetlabs.com/browse/RAZOR-158

Signed-off-by: Daniel Pittman <daniel@rimspace.net>