Allow JSON object syntax `name=val`
In order to remove JSON from the `create-broker` command's `--configuration` parameter, there needs to be an alternative way to construct the JSON object that's passed to the server. The user can now provide any argument as `--argument name=value` and the client will construct the JSON object passed to the server.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-198