Allow formatting of non-spec attributes
In cases like `razor nodes node1 tags`, navigation would previously reach the end of the navigation string and be left with just an array. With this array, it was not possible to customize the output (and was throwing an error).

This fix allows the navigation to retain the parent's spec string and store the navigation that was not consumed. It uses this by setting system attribute `+spec` inside the document. The view can then format the output as desired.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-242