Indicate available additional details
When the user ran `razor nodes node1`, it wasn't clear that the tree could be further traversed via e.g. `razor nodes node1 tags`. This adds a line at the end of the display to tell the next items that can be traversed.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-241