Razor should throw error when help is requested but does not exist
When a help page is requested but does not exist, e.g. `razor nodes --help`,
the command is either executed anyway or an inaccurate error message is thrown.
This case should instead respond with a message like: "Could not find help for
that entry".

Fixes https://tickets.puppetlabs.com/browse/RAZOR-375