RAZOR_API URL must start with http: or https:
Issues were arising from URLs that do not contain
the protocol at the start. This resulted in an
"opaque" message when trying to do navigation
and pass query parameters. URLs passed by -u
or --url will result in an error if they do not
start with the protocol.

For https://tickets.puppetlabs.com/browse/RAZOR-382