Use RAZOR_CA_FILE environment variable for cert verification
Communications with the razor server are being converted to SSL/TLS. In order
to verify that the Razor server is the correct server, a CA file can be
referenced. This change checks the RAZOR_CA_FILE path, which if supplied
must exist as a file, and uses that to check the authenticity of the razor
server.

For https://tickets.puppetlabs.com/browse/RAZOR-487