(maint) Have VCR run in isolated Bundler context for spec tests
When recording VCR entries for simulating server responses, the database was
being reset, which ran the `razor-admin` task. That task was failing since
the command was running inside the razor-client bundler which did not have the
proper gems loaded. This runs the same command in a clean environment outside
of bundler.

Another optimization is added here when recording, which is to only reset
the database if the current test uses VCR.