Add support for `aliases` syntax
A new property of argument metadata, `aliases`, has been added alongside the
existing `type` argument. This is an array that represents alternative names
for the given argument. The question is when to apply the alias, on the client
or on the server. The server can perform this reassignment too, but it's
simpler to have the client perform the reassignment as soon as possible, so
this commit does that.

For https://tickets.puppetlabs.com/browse/RAZOR-481