Fix HTTPS certificate mismatch on localhost
The razor server is now supplying certificate information to allow SSL
communications. By default, without any override, razor-client's URL to
talk to the razor server is https://localhost:8151/api. That's fine, except
that the server's certificate won't typically use "localhost" as its fqdn, so
the razor client will not be able to authenticate its own certificate.

This disables the verification in the case where the URL is localhost to
avoid this problem.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-529