Show command in event member
A new command has been added to razor-server that returns an event for
viewing. Previously, the 'command' property of this view was expecting an
entity. It is returned from the server as a string, indicating the URL for
the command that was executed. This attempts to display it as an entity
first, falling back to displaying whatever is there.

For https://tickets.puppetlabs.com/browse/RAZOR-388