Fit collection views to terminal width
If the user views a collection that has long widths for any of the columns, the
results would wrap in a manner that was very hard to read. This change takes
into account the size of the user's terminal window when rendering the table.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-571