Report razor-client version even if razor-server is unreachable
When the razor-client ran with `--version` and the Razor server was unreachable,
an error message was displayed and no client version was output. This is
annoying if the user just wants to check its client version.

The desired behavior, now implemented, is to display the client version
regardless, then show a connection error.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-627