Use the 'depth' parameter when querying collections
Previously, when the razor-client would query the
/api/collections/<type> endpoint via. `razor <type>`
(e.g. `razor nodes`), it would make O(N) requests to
get the full details of each item in the collection, where
N is the collection size. This does not scale well when N
is large.

The work in RAZOR-433 added a new 'depth' parameter to the
/api/collections/<type> endpoint that is either 0 or 1. 0
returns a list of item references, which is the original
default behavior, while 1 returns the full list of items.

The code changes in this commit consist of taking advantage
of the depth parameter. Specifically, if our final query is
to a collection endpoint that has a depth parameter, we set
the depth to 1 so that the endpoint returns the full item
details for us. This way, O(N) requests becomes O(1) requests,
i.e. the number of requests we make does not depend on N.
If the collection endpoint does not have a depth parameter,
then we don't use it -- the default behavior is preserved. This
enforces backwards compatibility with older razor-server versions.
Finally, we don't use the depth parameter if the collection endpoint
is an intermediate request (e.g. for something like `razor nodes <node_name>`)
because for that case, we will only query one item in the collection so
that object references are good enough, and better for performance.

Linked to https://tickets.puppetlabs.com/browse/RAZOR-1072