Pin `unf` to 0.1.4 (requiring native extensions)
Since `unf` version `0.2.0.beta2` hasn't been properly released at this
point, we will pin to 0.1.4 again. The trade-off here is between
requiring `--pre` for the prerelease version of the gem, and requiring
GCC to be installed for the released version of the gem. The latter is
what our most recent release follows, so we will stick with that.

We may wish to release `unf`'s `0.2.0.beta2` release out of band, which
will remove the need for building native extensions.

Fixes https://github.com/puppetlabs/RAZOR-1111